+++ source/Internal/Framework/Module/Configuration/Bridge/ModuleSettingBridgeInterface.php
@@ -9,6 +9,9 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Bridge;
 
+/**
+ * @deprecated use OxidEsales\EshopCommunity\Internal\Framework\Module\Facade\ModuleSettingServiceInterface
+ */
 interface ModuleSettingBridgeInterface
 {
     /**